Travis CI integration (resolves #1)
