+++ .travis.yaml
@@ -1,0 +1,3 @@
+install: pip install --editable .
+language: python
+script: pytest
