+++ tests/test_block.py
@@ -1,0 +1,18 @@
+def test_basic():
+    assert True
+
+
+def test_bottleneck():
+    assert True
+
+
+def test_convolution():
+    assert True
+
+
+def test_residual():
+    assert True
+
+
+def test_shortcut():
+    assert True
