+++ tests/test_model.py
@@ -1,0 +1,28 @@
+class TestResNet:
+    def test_constructor(self):
+        assert True
+
+
+class TestResNet18:
+    def test_constructor(self):
+        assert True
+
+
+class TestResNet34:
+    def test_constructor(self):
+        assert True
+
+
+class TestResNet50:
+    def test_constructor(self):
+        assert True
+
+
+class TestResNet101:
+    def test_constructor(self):
+        assert True
+
+
+class TestResNet152:
+    def test_constructor(self):
+        assert True
